feat/fix/build: Изменена версия сборки. Исправленны ошибки параметров запроса. Рефакторинг обработки получения ошибки http. Добавлен атрибут при перезаписи видео
