+++ build.gradle
@@ -2,11 +2,10 @@ plugins {
 	id 'java'
 	id 'org.springframework.boot' version '2.7.6'
 	id 'io.spring.dependency-management' version '1.0.15.RELEASE'
-	id 'org.bytedeco.gradle-javacpp-platform' version "1.5.10"
 }
 
 group = 'ru.nozdratenko'
-version = '2.5.10-SNAPSHOT'
+version = '2.5.11-SNAPSHOT'
 sourceCompatibility = '11'
 
 repositories {
@@ -17,10 +16,6 @@ repositories {
 	maven {
 		url 'https://www.dcm4che.org/maven2/'
 	}
-}
-
-ext {
-	javacppPlatform = 'linux-x86_64,windows-x86_64' // defaults to Loader.getPlatform()
 }
 
 dependencies {
