+++ src/main/frontend/src/helpers/alcometer.js
@@ -1,7 +1,5 @@
-import store from "@/store";
-import { ErrorCodes } from "vue";
-import { useToast } from "vue-toastification";
-const toast = useToast();
+import {defaultError} from "@/helpers/http-errors";
+
 
 export async function getAlcometerResult() {
     return await axios.post(`device/alcometer`).then(({ data }) => {            
@@ -32,21 +30,4 @@ async function changeMode(modeName) {
     return await axios.post(`device/alcometer/change-mode`, {mode: modeName}).then(({ data }) => {
         return data;
     }).catch(defaultError);
-}
-
-function defaultError(error) {
-    const data = error.response?.data;
-    if (data?.message) {
-        toast.error(data.message);
-    } else if (error.response) {
-        switch (error?.response?.status) {
-            case 400: toast.error('Ошибка авторизации запроса')
-            case 500: toast.error('Ошибка сервера')
-            default: toast.error('Неизвестная ошибка запроса')
-        }
-    } else {
-        store.$state.loseConnect = true;
-    }
-    
-    console.log(error);
 }
