+++ src/main/frontend/src/helpers/tonometer.js
@@ -1,6 +1,4 @@
-import store from "@/store";
-import { useToast } from "vue-toastification";
-const toast = useToast();
+import {defaultError} from "@/helpers/http-errors";
 
 export async function getPressure() {
     return await axios.post(`device/tonometer`).then(({ data }) => {            
@@ -28,20 +26,4 @@ export async function checkDevices() {
     }).catch((error) => {
         console.log(error);
     });
-}
-
-function defaultError(error) {
-    const data = error.response?.data;
-    if (data?.message) {
-        toast.error(data.message);
-    } else if(error.response) {
-        switch (error?.response?.status) {
-            case 400: toast.error('Ошибка авторизации запроса')
-            default: toast.error('Неизвестная ошибка запроса')
-        }
-    } else {
-        store.$state.loseConnect = true;
-    }
-    
-    console.log(error);
 }
