+++ src/main/java/ru/nozdratenko/sdpo/controller/Device/MediaController.java
@@ -33,9 +33,11 @@ public class MediaController {
 
     @PostMapping(value = "/device/media")
     @ResponseBody
-    public ResponseEntity media(@RequestBody Map<String, String> json, @RequestHeader Boolean restart) {
+    public ResponseEntity media(@RequestBody Map<String, String> json) {
         SdpoLog.info("make photo and video");
-        if(restart != null && restart) CameraHelper.restartMediaTask();
+        if(json.get("restart").equals("true")) {
+            CameraHelper.restartMediaTask();
+        }
         JSONObject result = CameraHelper.makePhotoAndVideo(json.get("driver_id"));
         return ResponseEntity.status(HttpStatus.OK).body(result.toMap());
     }
