+++ src/main/resources/static/index.html
@@ -1,1 +1,1 @@
-<!doctype html><html class="disable-mouse" lang=""><head><meta charset="utf-8"><meta http-equiv="X-UA-Compatible" content="IE=edge"><meta name="viewport" content="width=device-width,initial-scale=1"><link rel="icon" href="/favicon.ico"><title>СДПО Трансавто 7</title><script defer="defer" src="/js/chunk-vendors.06885a9c.js"></script><script defer="defer" src="/js/app.6c9f9186.js"></script><link href="/css/chunk-vendors.9e960663.css" rel="stylesheet"><link href="/css/app.eb0ad7bb.css" rel="stylesheet"></head><body><noscript><strong>Пожалуйста, включите javascript</strong></noscript><div id="app"></div></body></html>
+<!doctype html><html class="disable-mouse" lang=""><head><meta charset="utf-8"><meta http-equiv="X-UA-Compatible" content="IE=edge"><meta name="viewport" content="width=device-width,initial-scale=1"><link rel="icon" href="/favicon.ico"><title>СДПО Трансавто 7</title><script defer="defer" src="/js/chunk-vendors.06885a9c.js"></script><script defer="defer" src="/js/app.3e91c0e7.js"></script><link href="/css/chunk-vendors.9e960663.css" rel="stylesheet"><link href="/css/app.eb0ad7bb.css" rel="stylesheet"></head><body><noscript><strong>Пожалуйста, включите javascript</strong></noscript><div id="app"></div></body></html>
