--- src/main/resources/static/js/286.01f1e344.js
@@ -1,2 +1,0 @@
-"use strict";(self["webpackChunksdpo_frontend"]=self["webpackChunksdpo_frontend"]||[]).push([[286],{5611:function(t,e,n){n.d(e,{Xi:function(){return d},lm:function(){return c},qb:function(){return r},vX:function(){return o}});var s=n(7625),a=n(4677);const i=(0,a.pm)();async function o(){return await axios.post("device/tonometer").then((({data:t})=>t)).catch(u)}async function r(){return await axios.post("device/tonometer/disable").then((({data:t})=>t)).catch(u)}async function c(t="connect"){return await axios.post("device/tonometer/connect",{status:t}).then((({data:t})=>t)).catch(u)}async function d(){return await axios.post("device/scan").then((({data:t})=>t)).catch((t=>{console.log(t)}))}function u(t){const e=t.response?.data;if(e?.message)i.error(e.message);else if(t.response)switch(t?.response?.status){case 400:i.error("Ошибка авторизации запроса");default:i.error("Неизвестная ошибка запроса")}else s.Z.$state.loseConnect=!0;console.log(t)}},3286:function(t,e,n){n.r(e),n.d(e,{default:function(){return p}});n(7658);var s=n(3396);const a={class:"step-3__outer"},i=(0,s.uE)('<div class="step-3"><h3 class="animate__animated animate__fadeInDown">Измерение артериального давления</h3><div class="step-3__items"><div class="step-3__item animate__animated animate__fadeInDown d-1"><span>1</span> Закрепите манжету тонометра на левом плече на 2-3 см выше сгиба локтя </div><div class="step-3__item animate__animated animate__fadeInDown d-2"><span>2</span> Положите руку на стол </div><div class="step-3__item animate__animated animate__fadeInDown d-3"><span>3</span> Убедитесь, что манжета находится на уровне сердца </div><div class="step-3__item animate__animated animate__fadeInDown d-4"><span>4</span> Нажмите кнопку СТАРТ на тонометре </div></div><div class="step-3__subtext animate__animated animate__fadeInDown d-5"> Не двигайтесь, пока идет измерение.<br> После окончания измерения снимите манжету </div></div>',1),o={class:"step-buttons"};function r(t,e,n,r,c,d){return(0,s.wg)(),(0,s.iD)("div",a,[i,(0,s._)("div",o,[(0,s._)("button",{onClick:e[0]||(e[0]=e=>t.$router.push({name:"step-ride"})),class:"btn opacity blue"},"Назад"),JSON.parse(d.system.tonometer_skip)?((0,s.wg)(),(0,s.iD)("button",{key:0,onClick:e[1]||(e[1]=e=>t.$router.push({name:"step-thermometer"})),class:"btn"},"Продолжить")):(0,s.kq)("",!0)])])}var c=n(5611),d={data(){return{interval:null}},async mounted(){this.interval=setInterval((async()=>{const t=await(0,c.vX)();void 0!==t&&null!==t&&"next"!==t&&(t?.pulse&&(this.inspection.pulse=t.pulse),(t?.systolic||t?.diastolic)&&(this.inspection.tonometer=t.systolic+"/"+t.diastolic),clearInterval(this.interval),this.$router.push({name:"step-thermometer"}))}),1e3)},unmounted(){clearInterval(this.interval),(0,c.qb)()},computed:{inspection(){return this.$store.state.inspection},system(){return this.$store.state.config?.system||{}}}},u=n(89);const l=(0,u.Z)(d,[["render",r]]);var p=l}}]);
-//# sourceMappingURL=286.01f1e344.js.map
