+++ src/main/resources/static/js/286.a327899b.js
@@ -1,0 +1,2 @@
+"use strict";(self["webpackChunksdpo_frontend"]=self["webpackChunksdpo_frontend"]||[]).push([[286],{5611:function(t,n,e){e.d(n,{Xi:function(){return c},lm:function(){return o},qb:function(){return i},vX:function(){return s}});var a=e(1857);async function s(){return await axios.post("device/tonometer").then((({data:t})=>t)).catch(a.i)}async function i(){return await axios.post("device/tonometer/disable").then((({data:t})=>t)).catch(a.i)}async function o(t="connect"){return await axios.post("device/tonometer/connect",{status:t}).then((({data:t})=>t)).catch(a.i)}async function c(){return await axios.post("device/scan").then((({data:t})=>t)).catch((t=>{console.log(t)}))}},3286:function(t,n,e){e.r(n),e.d(n,{default:function(){return _}});e(7658);var a=e(3396);const s={class:"step-3__outer"},i=(0,a.uE)('<div class="step-3"><h3 class="animate__animated animate__fadeInDown">Измерение артериального давления</h3><div class="step-3__items"><div class="step-3__item animate__animated animate__fadeInDown d-1"><span>1</span> Закрепите манжету тонометра на левом плече на 2-3 см выше сгиба локтя </div><div class="step-3__item animate__animated animate__fadeInDown d-2"><span>2</span> Положите руку на стол </div><div class="step-3__item animate__animated animate__fadeInDown d-3"><span>3</span> Убедитесь, что манжета находится на уровне сердца </div><div class="step-3__item animate__animated animate__fadeInDown d-4"><span>4</span> Нажмите кнопку СТАРТ на тонометре </div></div><div class="step-3__subtext animate__animated animate__fadeInDown d-5"> Не двигайтесь, пока идет измерение.<br> После окончания измерения снимите манжету </div></div>',1),o={class:"step-buttons"};function c(t,n,e,c,r,d){return(0,a.wg)(),(0,a.iD)("div",s,[i,(0,a._)("div",o,[(0,a._)("button",{onClick:n[0]||(n[0]=n=>t.$router.push({name:"step-ride"})),class:"btn opacity blue"},"Назад"),JSON.parse(d.system.tonometer_skip)?((0,a.wg)(),(0,a.iD)("button",{key:0,onClick:n[1]||(n[1]=n=>t.$router.push({name:"step-thermometer"})),class:"btn"},"Продолжить")):(0,a.kq)("",!0)])])}var r=e(5611),d={data(){return{interval:null}},async mounted(){this.interval=setInterval((async()=>{const t=await(0,r.vX)();void 0!==t&&null!==t&&"next"!==t&&(t?.pulse&&(this.inspection.pulse=t.pulse),(t?.systolic||t?.diastolic)&&(this.inspection.tonometer=t.systolic+"/"+t.diastolic),clearInterval(this.interval),this.$router.push({name:"step-thermometer"}))}),1e3)},unmounted(){clearInterval(this.interval),(0,r.qb)()},computed:{inspection(){return this.$store.state.inspection},system(){return this.$store.state.config?.system||{}}}},u=e(89);const p=(0,u.Z)(d,[["render",c]]);var _=p}}]);
+//# sourceMappingURL=286.a327899b.js.map
