--- src/main/resources/static/js/381.7284e66c.js
@@ -1,2 +1,0 @@
-"use strict";(self["webpackChunksdpo_frontend"]=self["webpackChunksdpo_frontend"]||[]).push([[381],{2756:function(e,t,s){s.d(t,{i:function(){return o}});var n=s(7625),r=s(4677);const a=(0,r.pm)();async function o(){return await axios.post("device/thermometer").then((({data:e})=>e)).catch(i)}function i(e){const t=e.response?.data;if(t?.message)a.error(t.message);else if(e.response)switch(e?.response?.status){case 400:a.error("Ошибка авторизации запроса");case 500:a.error("Ошибка сервера");default:a.error("Неизвестная ошибка запроса")}else n.Z.$state.loseConnect=!0;console.log(e)}},3524:function(e,t,s){s.r(t),s.d(t,{default:function(){return d}});s(7658);var n=s(3396),r=s.p+"img/pirometer.732884f8.png";const a={class:"step-4__outer"},o=(0,n._)("div",{class:"step-4"},[(0,n._)("h3",{class:"animate__animated animate__fadeInUp"},"Измерение температуры тела"),(0,n._)("img",{class:"animate__animated animate__fadeInUp d-1",src:r})],-1),i={class:"step-buttons"};function c(e,t,s,r,c,u){return(0,n.wg)(),(0,n.iD)("div",a,[o,(0,n._)("div",i,[(0,n._)("button",{onClick:t[0]||(t[0]=t=>e.$router.push({name:"step-tonometer"})),class:"btn opacity blue"},"Назад"),JSON.parse(u.system.thermometer_skip)?((0,n.wg)(),(0,n.iD)("button",{key:0,onClick:t[1]||(t[1]=t=>e.$router.push({name:"step-alcometer"})),class:"btn"},"Продолжить")):(0,n.kq)("",!0)])])}var u=s(2756),l={data(){return{interval:null}},async mounted(){this.interval=setInterval((async()=>{const e=await(0,u.i)();void 0!==e&&null!==e&&"next"!==e&&(this.inspection.t_people=Number(e)||36.6,clearInterval(this.interval),this.$router.push({name:"step-alcometer"}))}),1e3)},unmounted(){clearInterval(this.interval)},computed:{inspection(){return this.$store.state.inspection},system(){return this.$store.state.config?.system||{}}}},p=s(89);const m=(0,p.Z)(l,[["render",c]]);var d=m}}]);
-//# sourceMappingURL=381.7284e66c.js.map
