+++ src/main/resources/static/js/381.acac5138.js
@@ -1,0 +1,2 @@
+"use strict";(self["webpackChunksdpo_frontend"]=self["webpackChunksdpo_frontend"]||[]).push([[381],{2756:function(t,e,n){n.d(e,{i:function(){return a}});var s=n(1857);async function a(){return await axios.post("device/thermometer").then((({data:t})=>t)).catch(s.i)}},3524:function(t,e,n){n.r(e),n.d(e,{default:function(){return d}});n(7658);var s=n(3396),a=n.p+"img/pirometer.732884f8.png";const r={class:"step-4__outer"},i=(0,s._)("div",{class:"step-4"},[(0,s._)("h3",{class:"animate__animated animate__fadeInUp"},"Измерение температуры тела"),(0,s._)("img",{class:"animate__animated animate__fadeInUp d-1",src:a})],-1),o={class:"step-buttons"};function c(t,e,n,a,c,u){return(0,s.wg)(),(0,s.iD)("div",r,[i,(0,s._)("div",o,[(0,s._)("button",{onClick:e[0]||(e[0]=e=>t.$router.push({name:"step-tonometer"})),class:"btn opacity blue"},"Назад"),JSON.parse(u.system.thermometer_skip)?((0,s.wg)(),(0,s.iD)("button",{key:0,onClick:e[1]||(e[1]=e=>t.$router.push({name:"step-alcometer"})),class:"btn"},"Продолжить")):(0,s.kq)("",!0)])])}var u=n(2756),l={data(){return{interval:null}},async mounted(){this.interval=setInterval((async()=>{const t=await(0,u.i)();void 0!==t&&null!==t&&"next"!==t&&(this.inspection.t_people=Number(t)||36.6,clearInterval(this.interval),this.$router.push({name:"step-alcometer"}))}),1e3)},unmounted(){clearInterval(this.interval)},computed:{inspection(){return this.$store.state.inspection},system(){return this.$store.state.config?.system||{}}}},p=n(89);const m=(0,p.Z)(l,[["render",c]]);var d=m}}]);
+//# sourceMappingURL=381.acac5138.js.map
