+++ src/main/resources/static/js/975.14e4eb48.js
@@ -1,0 +1,2 @@
+"use strict";(self["webpackChunksdpo_frontend"]=self["webpackChunksdpo_frontend"]||[]).push([[975],{2419:function(e,t,s){s.d(t,{NO:function(){return l},TQ:function(){return i},af:function(){return n},o:function(){return _}});var a=s(1857);async function n(){return await axios.post("device/photo").then((({data:e})=>e)).catch(a.i)}async function i(){return await axios.get("device/video/size").then((({data:e})=>e)).catch(a.i)}async function l(e,t=!1){return await axios.post("device/media",{driver_id:e,restart:t}).then((({data:e})=>e)).catch(a.i)}async function _(){return await axios.post("device/video/test").then((({data:e})=>e)).catch(a.i)}},2756:function(e,t,s){s.d(t,{i:function(){return n}});var a=s(1857);async function n(){return await axios.post("device/thermometer").then((({data:e})=>e)).catch(a.i)}},5611:function(e,t,s){s.d(t,{Xi:function(){return _},lm:function(){return l},qb:function(){return i},vX:function(){return n}});var a=s(1857);async function n(){return await axios.post("device/tonometer").then((({data:e})=>e)).catch(a.i)}async function i(){return await axios.post("device/tonometer/disable").then((({data:e})=>e)).catch(a.i)}async function l(e="connect"){return await axios.post("device/tonometer/connect",{status:e}).then((({data:e})=>e)).catch(a.i)}async function _(){return await axios.post("device/scan").then((({data:e})=>e)).catch((e=>{console.log(e)}))}},1975:function(e,t,s){s.r(t),s.d(t,{default:function(){return Cs}});var a=s(3396),n=s(7139);const i={class:"admin"},l={class:"admin__tabs"};function _(e,t,s,_,o,d){const c=(0,a.up)("testing"),m=(0,a.up)("tonometer"),r=(0,a.up)("password"),u=(0,a.up)("system"),p=(0,a.up)("logo");return(0,a.wg)(),(0,a.iD)("div",i,[(0,a._)("div",l,[(0,a._)("button",{class:(0,n.C_)(["admin__tab animate__animated animate__fadeInDown d-1",{active:"system"===o.selected}]),onClick:t[0]||(t[0]=e=>o.selected="system")}," Основные настройки ",2),(0,a._)("button",{class:(0,n.C_)(["admin__tab animate__animated animate__fadeInDown d-2",{active:"password"===o.selected}]),onClick:t[1]||(t[1]=e=>o.selected="password")}," Настройки пароля ",2),(0,a._)("button",{class:(0,n.C_)(["admin__tab animate__animated animate__fadeInDown d-3",{active:"tonometer"===o.selected}]),onClick:t[2]||(t[2]=e=>o.selected="tonometer")}," Настройки тонометра ",2),(0,a._)("button",{class:(0,n.C_)(["admin__tab animate__animated animate__fadeInDown d-3",{active:"logo"===o.selected}]),onClick:t[3]||(t[3]=e=>o.selected="logo")}," Смена логотипа ",2),(0,a._)("button",{class:(0,n.C_)(["admin__tab animate__animated animate__fadeInDown d-4",{active:"testing"===o.selected}]),onClick:t[4]||(t[4]=e=>o.selected="testing")}," Тестирование ",2)]),"testing"===o.selected?((0,a.wg)(),(0,a.j4)(c,{key:0})):(0,a.kq)("",!0),"tonometer"===o.selected?((0,a.wg)(),(0,a.j4)(m,{key:1})):(0,a.kq)("",!0),"password"===o.selected?((0,a.wg)(),(0,a.j4)(r,{key:2})):(0,a.kq)("",!0),"system"===o.selected?((0,a.wg)(),(0,a.j4)(u,{key:3})):(0,a.kq)("",!0),"logo"===o.selected?((0,a.wg)(),(0,a.j4)(p,{key:4})):(0,a.kq)("",!0)])}const o={class:"admin__testing"},d=["disabled"],c=["disabled"],m=["disabled"],r=["disabled"],u=["disabled"],p=["disabled"],y={key:0,class:"admin__loading animate__animated animate__fadeInUp"},v=(0,a._)("div",{class:"lds-ring"},[(0,a._)("div"),(0,a._)("div"),(0,a._)("div"),(0,a._)("div")],-1),h={key:1,class:"admin__testing-result"},w=["src"],b={key:1,class:"admin__testing-video animate__animated animate__fadeInUp"},g=(0,a._)("video",{autoplay:"autoplay",controls:""},[(0,a._)("source",{src:"http://localhost:8080/video",type:"video/mp4"})],-1),k=[g],f={key:2,class:"admin__testing-temp animate__animated animate__fadeInUp"},U=(0,a._)("span",null,"°C",-1),I={key:3,class:"admin__testing-temp animate__animated animate__fadeInUp"},C={key:4,class:"admin__testing-pressure animate__animated animate__fadeInUp"},x={class:"admin__testing-pressure-item"},D=(0,a._)("span",null,"Систолическое давление",-1),V={class:"admin__testing-pressure-item"},q=(0,a._)("span",null,"Диастолическое давление",-1),z={class:"admin__testing-pressure-item"},P=(0,a._)("span",null,"Пульс",-1);function $(e,t,s,i,l,_){return(0,a.wg)(),(0,a.iD)("div",o,[(0,a._)("button",{disabled:"loading"===l.show,onClick:t[0]||(t[0]=e=>_.photo()),class:"btn blue tab animate__animated animate__fadeInUp"},"Тестовый снимок",8,d),(0,a._)("button",{disabled:"loading"===l.show,onClick:t[1]||(t[1]=e=>_.video()),class:"btn blue tab animate__animated animate__fadeInUp d-1"},"Тестовое видео",8,c),(0,a._)("button",{disabled:"loading"===l.show,onClick:t[2]||(t[2]=e=>_.thermometer()),class:"btn blue tab animate__animated animate__fadeInUp d-2"},"Тест пирометра",8,m),(0,a._)("button",{disabled:"loading"===l.show,onClick:t[3]||(t[3]=e=>_.alcometer()),class:"btn blue tab animate__animated animate__fadeInUp d-3"},"Тест алкометра",8,r),(0,a._)("button",{disabled:"loading"===l.show,onClick:t[4]||(t[4]=e=>_.printer()),class:"btn blue tab animate__animated animate__fadeInUp d-4"},"Тестовая печать",8,u),(0,a._)("button",{disabled:"loading"===l.show,onClick:t[5]||(t[5]=e=>_.tonometer()),class:"btn blue tab animate__animated animate__fadeInUp d-5"},"Тест тонометра",8,p),"loading"===l.show?((0,a.wg)(),(0,a.iD)("div",y,[v,(0,a.Uk)(" Загрузка ")])):((0,a.wg)(),(0,a.iD)("div",h,["image"===l.show?((0,a.wg)(),(0,a.iD)("img",{key:0,class:"animate__animated animate__fadeInUp",src:l.image},null,8,w)):(0,a.kq)("",!0),"video"===l.show?((0,a.wg)(),(0,a.iD)("div",b,k)):(0,a.kq)("",!0),"temp"===l.show?((0,a.wg)(),(0,a.iD)("div",f,[(0,a.Uk)((0,n.zw)(l.temp)+" ",1),U])):(0,a.kq)("",!0),"alcometer"===l.show?((0,a.wg)(),(0,a.iD)("div",I,(0,n.zw)(void 0===l.alcometerPpm?"Не удалось получить результат":l.alcometerPpm+" ‰"),1)):(0,a.kq)("",!0),"pressure"===l.show?((0,a.wg)(),(0,a.iD)("div",C,[(0,a._)("div",x,[D,(0,a.Uk)(" "+(0,n.zw)(l.pressure?.systolic?l.pressure.systolic+" мм. рт.ст":"Результатов нет"),1)]),(0,a._)("div",V,[q,(0,a.Uk)(" "+(0,n.zw)(l.pressure?.diastolic?l.pressure.diastolic+" мм. рт.ст":"Результатов нет"),1)]),(0,a._)("div",z,[P,(0,a.Uk)(" "+(0,n.zw)(l.pressure?.pulse?l.pressure.pulse+" уд/мин":"Результатов нет"),1)])])):(0,a.kq)("",!0)]))])}var L=s(2419),S=s(5611),Z=s(2756),T=s(6700),j=s(4161),F=s(1857);async function X(){return await j.Z.post("device/printer").then((({data:e})=>e)).catch(F.i)}var A={data(){return{show:"",image:null,video:null,temp:null,alcometerPpm:null,pressure:null,timeout:0,loading:!1,interval:null}},methods:{async photo(){this.show="loading",clearInterval(this.interval),this.image=await(0,L.af)(),this.show="",this.image&&setTimeout((()=>{this.show="image"}),2e3)},async video(){this.show="loading",clearInterval(this.interval),this.video=await(0,L.o)(),this.show="",this.video&&(this.show="video")},async thermometer(){this.show="loading",clearInterval(this.interval),this.interval=setInterval((async()=>{this.temp=await(0,Z.i)(),"next"!==this.temp&&(this.show="temp",clearInterval(this.interval))}),1e3)},async alcometer(){this.show="loading",clearInterval(this.interval),this.interval=setInterval((async()=>{const e=await(0,T.pe)();"next"!==e&&(this.alcometerPpm=Number(e)||0,this.show="alcometer",clearInterval(this.interval))}),1e3)},async printer(){this.show="loading",this.alcometerPpm=await X(),this.show=""},async tonometer(){this.show="loading",clearInterval(this.interval),this.interval=setInterval((async()=>{const e=await(0,S.vX)();"next"!==e&&(this.pressure=e,clearInterval(this.interval),this.show="pressure")}),1e3)}}},H=s(89);const R=(0,H.Z)(A,[["render",$]]);var Y=R;const B={key:0,class:"admin__loading animate__animated animate__fadeInUp"},E=(0,a._)("div",{class:"lds-ring"},[(0,a._)("div"),(0,a._)("div"),(0,a._)("div"),(0,a._)("div")],-1),K={key:1,class:"admin__tonometer"},N={class:"admin__tonometer-card animate__animated animate__fadeInUp d-1"},Q=(0,a._)("div",{class:"admin__tonometer-card-title"}," Информация ",-1),M=["disabled"],O={class:"admin__tonometer-devices animate__animated animate__fadeInUp d-3"},W=["disabled","onClick"];function G(e,t,s,i,l,_){return l.result?((0,a.wg)(),(0,a.iD)("div",K,[(0,a._)("div",N,[Q,(0,a._)("p",null,"Подключенный тонометр: "+(0,n.zw)(_.config.main.tonometer_mac||"-"),1),(0,a._)("button",{disabled:l.connecting,class:(0,n.C_)([{red:l.connecting},"btn small blue"]),style:{width:"100%","margin-top":"10px"},onClick:t[0]||(t[0]=(...e)=>_.connect&&_.connect(...e))},"Подключить",10,M)]),(0,a._)("div",O,[((0,a.wg)(!0),(0,a.iD)(a.HY,null,(0,a.Ko)(l.result?.devices,(e=>((0,a.wg)(),(0,a.iD)("button",{class:"admin__tonometer-device",disabled:e.address===_.config.main.tonometer_mac,key:e.address,onClick:t=>_.save(e.address)},[(0,a.Uk)((0,n.zw)(e.address||"Неизвестный адрес")+" ",1),e.name?((0,a.wg)(),(0,a.iD)(a.HY,{key:0},[(0,a.Uk)(" - "+(0,n.zw)(e.name),1)],64)):(0,a.kq)("",!0)],8,W)))),128))])])):((0,a.wg)(),(0,a.iD)("div",B,[E,(0,a.Uk)(" Загрузка ")]))}var J=s(4457),ee=s(4677),te={data(){return{result:null,scan_task:null,connect_task:null,toast:(0,ee.pm)(),connecting:!1}},computed:{config(){return this.$store.state.config}},methods:{async getDevices(){this.scan_task=setInterval((async()=>{this.result=await(0,S.Xi)()}),1e3)},async save(e){await(0,J.Tc)(e),this.config.main.tonometer_mac=e,this.toast.success("Настройки тонометра сохранены")},async connect(){this.connecting=!0,this.connect_task=setInterval((async()=>{let e=await(0,S.lm)();"set"===e&&(this.connected=!0,this.connecting=!1,this.config.main.tonometer_connect=!0,clearInterval(this.connect_task))}),1e3)}},mounted(){this.getDevices()},unmounted(){clearInterval(this.scan_task),clearInterval(this.connect_task),(0,S.lm)("stop")}};const se=(0,H.Z)(te,[["render",G]]);var ae=se,ne=s(9242);const ie={class:"admin__password"},le={class:"password-form"},_e=(0,a._)("div",{class:"password-form__title animate__animated animate__fadeInDown"}," Введите новый пароль ",-1),oe={class:"password-form__input animate__animated animate__fadeIn d-1"},de={class:"number-buttons",ref:"numbers"},ce=(0,a._)("i",{class:"ri-close-fill"},null,-1),me=[ce],re=(0,a._)("i",{class:"ri-delete-back-line"},null,-1),ue=[re];function pe(e,t,s,n,i,l){return(0,a.wg)(),(0,a.iD)("div",ie,[(0,a._)("div",le,[_e,(0,a._)("div",oe,[(0,a.wy)((0,a._)("input",{type:"password","onUpdate:modelValue":t[0]||(t[0]=e=>i.password=e)},null,512),[[ne.nr,i.password]])]),(0,a._)("div",de,[(0,a._)("button",{class:"number-buttons__item",onClick:t[1]||(t[1]=e=>i.password+="1")},"1"),(0,a._)("button",{class:"number-buttons__item",onClick:t[2]||(t[2]=e=>i.password+="2")},"2"),(0,a._)("button",{class:"number-buttons__item",onClick:t[3]||(t[3]=e=>i.password+="3")},"3"),(0,a._)("button",{class:"number-buttons__item",onClick:t[4]||(t[4]=e=>i.password+="4")},"4"),(0,a._)("button",{class:"number-buttons__item",onClick:t[5]||(t[5]=e=>i.password+="5")},"5"),(0,a._)("button",{class:"number-buttons__item",onClick:t[6]||(t[6]=e=>i.password+="6")},"6"),(0,a._)("button",{class:"number-buttons__item",onClick:t[7]||(t[7]=e=>i.password+="7")},"7"),(0,a._)("button",{class:"number-buttons__item",onClick:t[8]||(t[8]=e=>i.password+="8")},"8"),(0,a._)("button",{class:"number-buttons__item",onClick:t[9]||(t[9]=e=>i.password+="9")},"9"),(0,a._)("button",{class:"number-buttons__item",onClick:t[10]||(t[10]=e=>i.password="")},me),(0,a._)("button",{class:"number-buttons__item",onClick:t[11]||(t[11]=e=>i.password+="0")},"0"),(0,a._)("button",{class:"number-buttons__item",onClick:t[12]||(t[12]=e=>i.password=i.password.slice(0,-1))},ue)],512),i.password.length>3?((0,a.wg)(),(0,a.iD)("button",{key:0,onClick:t[13]||(t[13]=(...e)=>l.save&&l.save(...e)),class:"btn animate__animated animate__fadeInUp"},"Сохранить")):(0,a.kq)("",!0)])])}var ye={data(){return{password:"",toast:(0,ee.pm)()}},methods:{async save(){await(0,J.vW)(this.password),this.password="",this.toast.success("Пароль успешно изменен"),this.$store.state.config.password=this.password}},mounted(){this.$refs.numbers.querySelectorAll("button").forEach(((e,t)=>{e.classList.add("animate__animated"),e.classList.add("animate__fadeInUp"),e.style.setProperty("animation-delay",.2+.05*t+"s")}))}};const ve=(0,H.Z)(ye,[["render",pe]]);var he=ve;const we={class:"admin__system"},be={class:"admin__system-card animate__animated animate__fadeInUp d-1"},ge=(0,a._)("div",{class:"admin__system-card__title"}," Основные настройки ",-1),ke={class:"admin__system-card__item"},fe=(0,a._)("span",null,"Информация о водителе",-1),Ue={class:"switch"},Ie=(0,a._)("div",{class:"slider round"},null,-1),Ce={class:"admin__system-card__item"},xe=(0,a._)("span",null,"Тип осмотра",-1),De={class:"switch"},Ve=(0,a._)("div",{class:"slider round"},null,-1),qe={class:"admin__system-card__item"},ze=(0,a._)("span",null,"Вопрос о сне",-1),Pe={class:"switch"},$e=(0,a._)("div",{class:"slider round"},null,-1),Le={class:"admin__system-card__item"},Se=(0,a._)("span",null,"вопрос о самочувствии",-1),Ze={class:"switch"},Te=(0,a._)("div",{class:"slider round"},null,-1),je={class:"admin__system-card__item"},Fe=(0,a._)("span",null,"Ручной режим",-1),Xe={class:"switch"},Ae=(0,a._)("div",{class:"slider round"},null,-1),He={class:"admin__system-card__item"},Re=(0,a._)("span",null,"Переход в начало",-1),Ye={class:"switch"},Be=(0,a._)("div",{class:"slider round"},null,-1),Ee={class:"admin__system-card__item"},Ke=(0,a._)("span",null,[(0,a.Uk)("Задержка перед "),(0,a._)("br"),(0,a.Uk)(" повторным прохождением "),(0,a._)("br"),(0,a.Uk)("(в миллисекундах)")],-1),Ne={class:"admin__system-card animate__animated animate__fadeInUp d-2"},Qe=(0,a._)("div",{class:"admin__system-card__title"}," Алкометр ",-1),Me={class:"admin__system-card__item"},Oe=(0,a._)("span",null,"Пропуск",-1),We={class:"switch"},Ge=(0,a._)("div",{class:"slider round"},null,-1),Je={class:"admin__system-card__item"},et=(0,a._)("span",null,"включен",-1),tt={class:"switch"},st=(0,a._)("div",{class:"slider round"},null,-1),at={class:"admin__system-card__item"},nt=(0,a._)("span",null,"Быстрый режим",-1),it={class:"switch"},lt=(0,a._)("div",{class:"slider round"},null,-1),_t={class:"admin__system-card__item"},ot=(0,a._)("span",null,"Количественный замер при положительном тесте",-1),dt={class:"switch"},ct=(0,a._)("div",{class:"slider round"},null,-1),mt={class:"admin__system-card animate__animated animate__fadeInUp d-3"},rt=(0,a._)("div",{class:"admin__system-card__title"}," Тонометр ",-1),ut={class:"admin__system-card__item"},pt=(0,a._)("span",null,"Пропуск",-1),yt={class:"switch"},vt=(0,a._)("div",{class:"slider round"},null,-1),ht={class:"admin__system-card__item"},wt=(0,a._)("span",null,"включен",-1),bt={class:"switch"},gt=(0,a._)("div",{class:"slider round"},null,-1),kt={class:"admin__system-card animate__animated animate__fadeInUp d-4"},ft=(0,a._)("div",{class:"admin__system-card__title"}," Камера ",-1),Ut={class:"admin__system-card__item"},It=(0,a._)("span",null,"Качество",-1),Ct=["value"],xt={class:"admin__system-card__item"},Dt=(0,a._)("span",null,"Видео",-1),Vt={class:"switch"},qt=(0,a._)("div",{class:"slider round"},null,-1),zt={class:"admin__system-card__item"},Pt=(0,a._)("span",null,"Фото",-1),$t={class:"switch"},Lt=(0,a._)("div",{class:"slider round"},null,-1),St={class:"admin__system-card__item"},Zt=(0,a._)("span",null,"Фото водителя",-1),Tt={class:"switch"},jt=(0,a._)("div",{class:"slider round"},null,-1),Ft={class:"admin__system-card animate__animated animate__fadeInUp d-5"},Xt=(0,a._)("div",{class:"admin__system-card__title"}," Принтер ",-1),At={class:"admin__system-card__item"},Ht=(0,a._)("span",null,"Печать",-1),Rt={class:"switch"},Yt=(0,a._)("div",{class:"slider round"},null,-1),Bt={class:"admin__system-card__item"},Et=(0,a._)("span",null,"Количество",-1),Kt={class:"admin__system-card animate__animated animate__fadeInUp d-6"},Nt=(0,a._)("div",{class:"admin__system-card__title"}," Термометр ",-1),Qt={class:"admin__system-card__item"},Mt=(0,a._)("span",null,"Пропуск",-1),Ot={class:"switch"},Wt=(0,a._)("div",{class:"slider round"},null,-1),Gt={class:"admin__system-card__item"},Jt=(0,a._)("span",null,"включен",-1),es={class:"switch"},ts=(0,a._)("div",{class:"slider round"},null,-1),ss={class:"admin__system-card block animate__animated animate__fadeInUp d-6"},as={class:"admin__system-card__title"},ns={class:"admin__system-card__item"},is=(0,a._)("span",null,"Адрес",-1),ls={class:"admin__system-card__item"},_s=(0,a._)("span",null,"Токен",-1),os={class:"admin__system-footer animate__animated animate__fadeInUp d-7"};function ds(e,t,s,i,l,_){return(0,a.wg)(),(0,a.iD)("div",we,[(0,a._)("div",be,[ge,(0,a._)("div",ke,[fe,(0,a._)("label",Ue,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[0]||(t[0]=e=>_.system.driver_info=e)},null,512),[[ne.e8,_.system.driver_info]]),Ie])]),(0,a._)("div",Ce,[xe,(0,a._)("label",De,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[1]||(t[1]=e=>_.system.type_ride=e)},null,512),[[ne.e8,_.system.type_ride]]),Ve])]),(0,a._)("div",qe,[ze,(0,a._)("label",Pe,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[2]||(t[2]=e=>_.system.question_sleep=e)},null,512),[[ne.e8,_.system.question_sleep]]),$e])]),(0,a._)("div",Le,[Se,(0,a._)("label",Ze,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[3]||(t[3]=e=>_.system.question_helth=e)},null,512),[[ne.e8,_.system.question_helth]]),Te])]),(0,a._)("div",je,[Fe,(0,a._)("label",Xe,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[4]||(t[4]=e=>_.system.manual_mode=e)},null,512),[[ne.e8,_.system.manual_mode]]),Ae])]),(0,a._)("div",He,[Re,(0,a._)("label",Ye,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[5]||(t[5]=e=>_.system.auto_start=e)},null,512),[[ne.e8,_.system.auto_start]]),Be])]),(0,a._)("div",Ee,[Ke,(0,a.wy)((0,a._)("input",{min:"1000",class:"medium",type:"number","onUpdate:modelValue":t[6]||(t[6]=e=>_.system.delay_before_retry_inspection=e)},null,512),[[ne.nr,_.system.delay_before_retry_inspection]])])]),(0,a._)("div",Ne,[Qe,(0,a._)("div",Me,[Oe,(0,a._)("label",We,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[7]||(t[7]=e=>_.system.alcometer_skip=e)},null,512),[[ne.e8,_.system.alcometer_skip]]),Ge])]),(0,a._)("div",Je,[et,(0,a._)("label",tt,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[8]||(t[8]=e=>_.system.alcometer_visible=e)},null,512),[[ne.e8,_.system.alcometer_visible]]),st])]),(0,a._)("div",at,[nt,(0,a._)("label",it,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[9]||(t[9]=e=>_.system.alcometer_fast=e)},null,512),[[ne.e8,_.system.alcometer_fast]]),lt])]),(0,a._)("div",_t,[ot,(0,a._)("label",dt,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[10]||(t[10]=e=>_.system.alcometer_retry=e)},null,512),[[ne.e8,_.system.alcometer_retry]]),ct])])]),(0,a._)("div",mt,[rt,(0,a._)("div",ut,[pt,(0,a._)("label",yt,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[11]||(t[11]=e=>_.system.tonometer_skip=e)},null,512),[[ne.e8,_.system.tonometer_skip]]),vt])]),(0,a._)("div",ht,[wt,(0,a._)("label",bt,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[12]||(t[12]=e=>_.system.tonometer_visible=e)},null,512),[[ne.e8,_.system.tonometer_visible]]),gt])])]),(0,a._)("div",kt,[ft,(0,a._)("div",Ut,[It,(0,a.wy)((0,a._)("select",{"onUpdate:modelValue":t[13]||(t[13]=e=>_.system.camera_dimension=e)},[((0,a.wg)(!0),(0,a.iD)(a.HY,null,(0,a.Ko)(l.videoSizes,(e=>((0,a.wg)(),(0,a.iD)("option",{key:e,value:e},(0,n.zw)(e),9,Ct)))),128))],512),[[ne.bM,_.system.camera_dimension]])]),(0,a._)("div",xt,[Dt,(0,a._)("label",Vt,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[14]||(t[14]=e=>_.system.camera_video=e)},null,512),[[ne.e8,_.system.camera_video]]),qt])]),(0,a._)("div",zt,[Pt,(0,a._)("label",$t,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[15]||(t[15]=e=>_.system.camera_photo=e)},null,512),[[ne.e8,_.system.camera_photo]]),Lt])]),(0,a._)("div",St,[Zt,(0,a._)("label",Tt,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[16]||(t[16]=e=>_.system.driver_photo=e)},null,512),[[ne.e8,_.system.driver_photo]]),jt])])]),(0,a._)("div",Ft,[Xt,(0,a._)("div",At,[Ht,(0,a._)("label",Rt,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[17]||(t[17]=e=>_.system.printer_write=e)},null,512),[[ne.e8,_.system.printer_write]]),Yt])]),(0,a._)("div",Bt,[Et,(0,a.wy)((0,a._)("input",{min:"1",class:"small",type:"number","onUpdate:modelValue":t[18]||(t[18]=e=>_.system.print_count=e)},null,512),[[ne.nr,_.system.print_count]])])]),(0,a._)("div",Kt,[Nt,(0,a._)("div",Qt,[Mt,(0,a._)("label",Ot,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[19]||(t[19]=e=>_.system.thermometer_skip=e)},null,512),[[ne.e8,_.system.thermometer_skip]]),Wt])]),(0,a._)("div",Gt,[Jt,(0,a._)("label",es,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":t[20]||(t[20]=e=>_.system.thermometer_visible=e)},null,512),[[ne.e8,_.system.thermometer_visible]]),ts])])]),(0,a._)("div",ss,[(0,a._)("div",as,[(0,a.Uk)(" Настройки API "),(0,a._)("div",{class:(0,n.C_)(["connection-status",{active:_.connection}])},[(0,a._)("span",null,(0,n.zw)(_.connection),1)],2)]),(0,a._)("div",ns,[is,(0,a.wy)((0,a._)("input",{"onUpdate:modelValue":t[21]||(t[21]=e=>_.config.url=e)},null,512),[[ne.nr,_.config.url]])]),(0,a._)("div",ls,[_s,(0,a.wy)((0,a._)("input",{"onUpdate:modelValue":t[22]||(t[22]=e=>_.config.token=e)},null,512),[[ne.nr,_.config.token]])])]),(0,a._)("div",os,[(0,a._)("button",{onClick:t[23]||(t[23]=(...e)=>_.save&&_.save(...e)),class:"btn blue"},"Сохранить")])])}var cs=s(6509),ms={data(){return{toast:(0,ee.pm)(),videoSizes:[]}},async mounted(){this.videoSizes=await(0,L.TQ)()},methods:{async save(){await(0,J.hR)(this.system),await(0,J.XP)(this.config.url,this.config.token),this.$store.state.point=await(0,cs.F8)(),this.toast.success("Настройки сохранены")}},computed:{system(){return this.$store.state.config?.system||{}},config(){return this.$store.state.config?.main||{}},connection(){return this.$store.state.connection||!1}}};const rs=(0,H.Z)(ms,[["render",ds]]);var us=rs;const ps={class:"admin__logo"},ys={class:"input-file"},vs=(0,a._)("span",null,"Выберите файл",-1),hs={key:0,class:"preview-image"},ws=["src"];function bs(e,t,s,n,i,l){return(0,a.wg)(),(0,a.iD)("div",ps,[(0,a._)("label",ys,[(0,a._)("input",{type:"file",onChange:t[0]||(t[0]=(...e)=>l.onFileChange&&l.onFileChange(...e))},null,32),vs]),l.logo?((0,a.wg)(),(0,a.iD)("div",hs,[(0,a._)("img",{src:l.logo},null,8,ws),(0,a._)("button",{onClick:t[1]||(t[1]=(...e)=>l.deleteLogo&&l.deleteLogo(...e)),class:"btn small"},"Удалить")])):(0,a.kq)("",!0)])}var gs={data(){return{toast:(0,ee.pm)()}},methods:{async onFileChange(e){const t=e.target.files[0],s=await this.getBase64(t);this.$store.state.config.main.logo=s,await(0,cs.Sx)(s),this.toast.success("Логотип успешно установлен")},getBase64(e){return new Promise(((t,s)=>{const a=new FileReader;a.readAsDataURL(e),a.onload=()=>{t(a.result)},a.onerror=e=>{console.log("Error: ",e)}}))},async deleteLogo(){this.$store.state.config.main.logo="",await(0,cs.Sx)(""),this.toast.success("Логотип успешно удален")}},computed:{logo(){return this.$store.state.config?.main?.logo}}};const ks=(0,H.Z)(gs,[["render",bs]]);var fs=ks,Us={components:{Testing:Y,Tonometer:ae,Password:he,System:us,Logo:fs},data(){return{selected:"system"}}};const Is=(0,H.Z)(Us,[["render",_]]);var Cs=Is}}]);
+//# sourceMappingURL=975.14e4eb48.js.map
