--- src/main/resources/static/js/975.faf22561.js
@@ -1,2 +1,0 @@
-"use strict";(self["webpackChunksdpo_frontend"]=self["webpackChunksdpo_frontend"]||[]).push([[975],{2419:function(e,s,t){t.d(s,{NO:function(){return o},TQ:function(){return l},af:function(){return i},o:function(){return _}});var a=t(7625),n=t(4677);(0,n.pm)();async function i(){return await axios.post("device/photo").then((({data:e})=>e)).catch(d)}async function l(){return await axios.get("device/video/size").then((({data:e})=>e)).catch(d)}async function o(e){return await axios.post("device/media",{driver_id:e}).then((({data:e})=>e)).catch(d)}async function _(){return await axios.post("device/video/test").then((({data:e})=>e)).catch(d)}function d(e){e.response&&(a.Z.$state.loseConnect=!0),console.log(e)}},2756:function(e,s,t){t.d(s,{i:function(){return l}});var a=t(7625),n=t(4677);const i=(0,n.pm)();async function l(){return await axios.post("device/thermometer").then((({data:e})=>e)).catch(o)}function o(e){const s=e.response?.data;if(s?.message)i.error(s.message);else if(e.response)switch(e?.response?.status){case 400:i.error("Ошибка авторизации запроса");case 500:i.error("Ошибка сервера");default:i.error("Неизвестная ошибка запроса")}else a.Z.$state.loseConnect=!0;console.log(e)}},5611:function(e,s,t){t.d(s,{Xi:function(){return d},lm:function(){return _},qb:function(){return o},vX:function(){return l}});var a=t(7625),n=t(4677);const i=(0,n.pm)();async function l(){return await axios.post("device/tonometer").then((({data:e})=>e)).catch(c)}async function o(){return await axios.post("device/tonometer/disable").then((({data:e})=>e)).catch(c)}async function _(e="connect"){return await axios.post("device/tonometer/connect",{status:e}).then((({data:e})=>e)).catch(c)}async function d(){return await axios.post("device/scan").then((({data:e})=>e)).catch((e=>{console.log(e)}))}function c(e){const s=e.response?.data;if(s?.message)i.error(s.message);else if(e.response)switch(e?.response?.status){case 400:i.error("Ошибка авторизации запроса");default:i.error("Неизвестная ошибка запроса")}else a.Z.$state.loseConnect=!0;console.log(e)}},1975:function(e,s,t){t.r(s),t.d(s,{default:function(){return xt}});var a=t(3396),n=t(7139);const i={class:"admin"},l={class:"admin__tabs"};function o(e,s,t,o,_,d){const c=(0,a.up)("testing"),r=(0,a.up)("tonometer"),m=(0,a.up)("password"),u=(0,a.up)("system"),p=(0,a.up)("logo");return(0,a.wg)(),(0,a.iD)("div",i,[(0,a._)("div",l,[(0,a._)("button",{class:(0,n.C_)(["admin__tab animate__animated animate__fadeInDown d-1",{active:"system"===_.selected}]),onClick:s[0]||(s[0]=e=>_.selected="system")}," Основные настройки ",2),(0,a._)("button",{class:(0,n.C_)(["admin__tab animate__animated animate__fadeInDown d-2",{active:"password"===_.selected}]),onClick:s[1]||(s[1]=e=>_.selected="password")}," Настройки пароля ",2),(0,a._)("button",{class:(0,n.C_)(["admin__tab animate__animated animate__fadeInDown d-3",{active:"tonometer"===_.selected}]),onClick:s[2]||(s[2]=e=>_.selected="tonometer")}," Настройки тонометра ",2),(0,a._)("button",{class:(0,n.C_)(["admin__tab animate__animated animate__fadeInDown d-3",{active:"logo"===_.selected}]),onClick:s[3]||(s[3]=e=>_.selected="logo")}," Смена логотипа ",2),(0,a._)("button",{class:(0,n.C_)(["admin__tab animate__animated animate__fadeInDown d-4",{active:"testing"===_.selected}]),onClick:s[4]||(s[4]=e=>_.selected="testing")}," Тестирование ",2)]),"testing"===_.selected?((0,a.wg)(),(0,a.j4)(c,{key:0})):(0,a.kq)("",!0),"tonometer"===_.selected?((0,a.wg)(),(0,a.j4)(r,{key:1})):(0,a.kq)("",!0),"password"===_.selected?((0,a.wg)(),(0,a.j4)(m,{key:2})):(0,a.kq)("",!0),"system"===_.selected?((0,a.wg)(),(0,a.j4)(u,{key:3})):(0,a.kq)("",!0),"logo"===_.selected?((0,a.wg)(),(0,a.j4)(p,{key:4})):(0,a.kq)("",!0)])}const _={class:"admin__testing"},d=["disabled"],c=["disabled"],r=["disabled"],m=["disabled"],u=["disabled"],p=["disabled"],y={key:0,class:"admin__loading animate__animated animate__fadeInUp"},v=(0,a._)("div",{class:"lds-ring"},[(0,a._)("div"),(0,a._)("div"),(0,a._)("div"),(0,a._)("div")],-1),h={key:1,class:"admin__testing-result"},w=["src"],b={key:1,class:"admin__testing-video animate__animated animate__fadeInUp"},g=(0,a._)("video",{autoplay:"autoplay",controls:""},[(0,a._)("source",{src:"http://localhost:8080/video",type:"video/mp4"})],-1),f=[g],k={key:2,class:"admin__testing-temp animate__animated animate__fadeInUp"},U=(0,a._)("span",null,"°C",-1),I={key:3,class:"admin__testing-temp animate__animated animate__fadeInUp"},C={key:4,class:"admin__testing-pressure animate__animated animate__fadeInUp"},x={class:"admin__testing-pressure-item"},D=(0,a._)("span",null,"Систолическое давление",-1),V={class:"admin__testing-pressure-item"},q=(0,a._)("span",null,"Диастолическое давление",-1),z={class:"admin__testing-pressure-item"},$=(0,a._)("span",null,"Пульс",-1);function Z(e,s,t,i,l,o){return(0,a.wg)(),(0,a.iD)("div",_,[(0,a._)("button",{disabled:"loading"===l.show,onClick:s[0]||(s[0]=e=>o.photo()),class:"btn blue tab animate__animated animate__fadeInUp"},"Тестовый снимок",8,d),(0,a._)("button",{disabled:"loading"===l.show,onClick:s[1]||(s[1]=e=>o.video()),class:"btn blue tab animate__animated animate__fadeInUp d-1"},"Тестовое видео",8,c),(0,a._)("button",{disabled:"loading"===l.show,onClick:s[2]||(s[2]=e=>o.thermometer()),class:"btn blue tab animate__animated animate__fadeInUp d-2"},"Тест пирометра",8,r),(0,a._)("button",{disabled:"loading"===l.show,onClick:s[3]||(s[3]=e=>o.alcometer()),class:"btn blue tab animate__animated animate__fadeInUp d-3"},"Тест алкометра",8,m),(0,a._)("button",{disabled:"loading"===l.show,onClick:s[4]||(s[4]=e=>o.printer()),class:"btn blue tab animate__animated animate__fadeInUp d-4"},"Тестовая печать",8,u),(0,a._)("button",{disabled:"loading"===l.show,onClick:s[5]||(s[5]=e=>o.tonometer()),class:"btn blue tab animate__animated animate__fadeInUp d-5"},"Тест тонометра",8,p),"loading"===l.show?((0,a.wg)(),(0,a.iD)("div",y,[v,(0,a.Uk)(" Загрузка ")])):((0,a.wg)(),(0,a.iD)("div",h,["image"===l.show?((0,a.wg)(),(0,a.iD)("img",{key:0,class:"animate__animated animate__fadeInUp",src:l.image},null,8,w)):(0,a.kq)("",!0),"video"===l.show?((0,a.wg)(),(0,a.iD)("div",b,f)):(0,a.kq)("",!0),"temp"===l.show?((0,a.wg)(),(0,a.iD)("div",k,[(0,a.Uk)((0,n.zw)(l.temp)+" ",1),U])):(0,a.kq)("",!0),"alcometer"===l.show?((0,a.wg)(),(0,a.iD)("div",I,(0,n.zw)(void 0===l.alcometerPpm?"Не удалось получить результат":l.alcometerPpm+" ‰"),1)):(0,a.kq)("",!0),"pressure"===l.show?((0,a.wg)(),(0,a.iD)("div",C,[(0,a._)("div",x,[D,(0,a.Uk)(" "+(0,n.zw)(l.pressure?.systolic?l.pressure.systolic+" мм. рт.ст":"Результатов нет"),1)]),(0,a._)("div",V,[q,(0,a.Uk)(" "+(0,n.zw)(l.pressure?.diastolic?l.pressure.diastolic+" мм. рт.ст":"Результатов нет"),1)]),(0,a._)("div",z,[$,(0,a.Uk)(" "+(0,n.zw)(l.pressure?.pulse?l.pressure.pulse+" уд/мин":"Результатов нет"),1)])])):(0,a.kq)("",!0)]))])}var P=t(2419),L=t(5611),S=t(2756),T=t(6700),j=t(7625),F=t(4161);async function X(){return await F.Z.post("device/printer").then((({data:e})=>e)).catch(A)}function A(e){const s=e.response?.data;if(s?.message)toast.error(s.message);else if(s.response)switch(e?.response?.status){case 400:toast.error("Ошибка авторизации запроса");default:toast.error("Неизвестная ошибка запроса")}else j.Z.$state.loseConnect=!0;console.log(e)}var H={data(){return{show:"",image:null,video:null,temp:null,alcometerPpm:null,pressure:null,timeout:0,loading:!1,interval:null}},methods:{async photo(){this.show="loading",clearInterval(this.interval),this.image=await(0,P.af)(),this.show="",this.image&&setTimeout((()=>{this.show="image"}),2e3)},async video(){this.show="loading",clearInterval(this.interval),this.video=await(0,P.o)(),this.show="",this.video&&(this.show="video")},async thermometer(){this.show="loading",clearInterval(this.interval),this.interval=setInterval((async()=>{this.temp=await(0,S.i)(),"next"!==this.temp&&(this.show="temp",clearInterval(this.interval))}),1e3)},async alcometer(){this.show="loading",clearInterval(this.interval),this.interval=setInterval((async()=>{const e=await(0,T.pe)();"next"!==e&&(this.alcometerPpm=Number(e)||0,this.show="alcometer",clearInterval(this.interval))}),1e3)},async printer(){this.show="loading",this.alcometerPpm=await X(),this.show=""},async tonometer(){this.show="loading",clearInterval(this.interval),this.interval=setInterval((async()=>{const e=await(0,L.vX)();"next"!==e&&(this.pressure=e,clearInterval(this.interval),this.show="pressure")}),1e3)}}},R=t(89);const Y=(0,R.Z)(H,[["render",Z]]);var B=Y;const E={key:0,class:"admin__loading animate__animated animate__fadeInUp"},K=(0,a._)("div",{class:"lds-ring"},[(0,a._)("div"),(0,a._)("div"),(0,a._)("div"),(0,a._)("div")],-1),N={key:1,class:"admin__tonometer"},Q={class:"admin__tonometer-card animate__animated animate__fadeInUp d-1"},M=(0,a._)("div",{class:"admin__tonometer-card-title"}," Информация ",-1),O=["disabled"],W={class:"admin__tonometer-devices animate__animated animate__fadeInUp d-3"},G=["disabled","onClick"];function J(e,s,t,i,l,o){return l.result?((0,a.wg)(),(0,a.iD)("div",N,[(0,a._)("div",Q,[M,(0,a._)("p",null,"Подключенный тонометр: "+(0,n.zw)(o.config.main.tonometer_mac||"-"),1),(0,a._)("button",{disabled:l.connecting,class:(0,n.C_)([{red:l.connecting},"btn small blue"]),style:{width:"100%","margin-top":"10px"},onClick:s[0]||(s[0]=(...e)=>o.connect&&o.connect(...e))},"Подключить",10,O)]),(0,a._)("div",W,[((0,a.wg)(!0),(0,a.iD)(a.HY,null,(0,a.Ko)(l.result?.devices,(e=>((0,a.wg)(),(0,a.iD)("button",{class:"admin__tonometer-device",disabled:e.address===o.config.main.tonometer_mac,key:e.address,onClick:s=>o.save(e.address)},[(0,a.Uk)((0,n.zw)(e.address||"Неизвестный адрес")+" ",1),e.name?((0,a.wg)(),(0,a.iD)(a.HY,{key:0},[(0,a.Uk)(" - "+(0,n.zw)(e.name),1)],64)):(0,a.kq)("",!0)],8,G)))),128))])])):((0,a.wg)(),(0,a.iD)("div",E,[K,(0,a.Uk)(" Загрузка ")]))}var ee=t(4457),se=t(4677),te={data(){return{result:null,scan_task:null,connect_task:null,toast:(0,se.pm)(),connecting:!1}},computed:{config(){return this.$store.state.config}},methods:{async getDevices(){this.scan_task=setInterval((async()=>{this.result=await(0,L.Xi)()}),1e3)},async save(e){await(0,ee.Tc)(e),this.config.main.tonometer_mac=e,this.toast.success("Настройки тонометра сохранены")},async connect(){this.connecting=!0,this.connect_task=setInterval((async()=>{let e=await(0,L.lm)();"set"===e&&(this.connected=!0,this.connecting=!1,this.config.main.tonometer_connect=!0,clearInterval(this.connect_task))}),1e3)}},mounted(){this.getDevices()},unmounted(){clearInterval(this.scan_task),clearInterval(this.connect_task),(0,L.lm)("stop")}};const ae=(0,R.Z)(te,[["render",J]]);var ne=ae,ie=t(9242);const le={class:"admin__password"},oe={class:"password-form"},_e=(0,a._)("div",{class:"password-form__title animate__animated animate__fadeInDown"}," Введите новый пароль ",-1),de={class:"password-form__input animate__animated animate__fadeIn d-1"},ce={class:"number-buttons",ref:"numbers"},re=(0,a._)("i",{class:"ri-close-fill"},null,-1),me=[re],ue=(0,a._)("i",{class:"ri-delete-back-line"},null,-1),pe=[ue];function ye(e,s,t,n,i,l){return(0,a.wg)(),(0,a.iD)("div",le,[(0,a._)("div",oe,[_e,(0,a._)("div",de,[(0,a.wy)((0,a._)("input",{type:"password","onUpdate:modelValue":s[0]||(s[0]=e=>i.password=e)},null,512),[[ie.nr,i.password]])]),(0,a._)("div",ce,[(0,a._)("button",{class:"number-buttons__item",onClick:s[1]||(s[1]=e=>i.password+="1")},"1"),(0,a._)("button",{class:"number-buttons__item",onClick:s[2]||(s[2]=e=>i.password+="2")},"2"),(0,a._)("button",{class:"number-buttons__item",onClick:s[3]||(s[3]=e=>i.password+="3")},"3"),(0,a._)("button",{class:"number-buttons__item",onClick:s[4]||(s[4]=e=>i.password+="4")},"4"),(0,a._)("button",{class:"number-buttons__item",onClick:s[5]||(s[5]=e=>i.password+="5")},"5"),(0,a._)("button",{class:"number-buttons__item",onClick:s[6]||(s[6]=e=>i.password+="6")},"6"),(0,a._)("button",{class:"number-buttons__item",onClick:s[7]||(s[7]=e=>i.password+="7")},"7"),(0,a._)("button",{class:"number-buttons__item",onClick:s[8]||(s[8]=e=>i.password+="8")},"8"),(0,a._)("button",{class:"number-buttons__item",onClick:s[9]||(s[9]=e=>i.password+="9")},"9"),(0,a._)("button",{class:"number-buttons__item",onClick:s[10]||(s[10]=e=>i.password="")},me),(0,a._)("button",{class:"number-buttons__item",onClick:s[11]||(s[11]=e=>i.password+="0")},"0"),(0,a._)("button",{class:"number-buttons__item",onClick:s[12]||(s[12]=e=>i.password=i.password.slice(0,-1))},pe)],512),i.password.length>3?((0,a.wg)(),(0,a.iD)("button",{key:0,onClick:s[13]||(s[13]=(...e)=>l.save&&l.save(...e)),class:"btn animate__animated animate__fadeInUp"},"Сохранить")):(0,a.kq)("",!0)])])}var ve={data(){return{password:"",toast:(0,se.pm)()}},methods:{async save(){await(0,ee.vW)(this.password),this.password="",this.toast.success("Пароль успешно изменен"),this.$store.state.config.password=this.password}},mounted(){this.$refs.numbers.querySelectorAll("button").forEach(((e,s)=>{e.classList.add("animate__animated"),e.classList.add("animate__fadeInUp"),e.style.setProperty("animation-delay",.2+.05*s+"s")}))}};const he=(0,R.Z)(ve,[["render",ye]]);var we=he;const be={class:"admin__system"},ge={class:"admin__system-card animate__animated animate__fadeInUp d-1"},fe=(0,a._)("div",{class:"admin__system-card__title"}," Основные настройки ",-1),ke={class:"admin__system-card__item"},Ue=(0,a._)("span",null,"Информация о водителе",-1),Ie={class:"switch"},Ce=(0,a._)("div",{class:"slider round"},null,-1),xe={class:"admin__system-card__item"},De=(0,a._)("span",null,"Тип осмотра",-1),Ve={class:"switch"},qe=(0,a._)("div",{class:"slider round"},null,-1),ze={class:"admin__system-card__item"},$e=(0,a._)("span",null,"Вопрос о сне",-1),Ze={class:"switch"},Pe=(0,a._)("div",{class:"slider round"},null,-1),Le={class:"admin__system-card__item"},Se=(0,a._)("span",null,"вопрос о самочувствии",-1),Te={class:"switch"},je=(0,a._)("div",{class:"slider round"},null,-1),Fe={class:"admin__system-card__item"},Xe=(0,a._)("span",null,"Ручной режим",-1),Ae={class:"switch"},He=(0,a._)("div",{class:"slider round"},null,-1),Re={class:"admin__system-card__item"},Ye=(0,a._)("span",null,"Переход в начало",-1),Be={class:"switch"},Ee=(0,a._)("div",{class:"slider round"},null,-1),Ke={class:"admin__system-card__item"},Ne=(0,a._)("span",null,[(0,a.Uk)("Задержка перед "),(0,a._)("br"),(0,a.Uk)(" повторным прохождением "),(0,a._)("br"),(0,a.Uk)("(в миллисекундах)")],-1),Qe={class:"admin__system-card animate__animated animate__fadeInUp d-2"},Me=(0,a._)("div",{class:"admin__system-card__title"}," Алкометр ",-1),Oe={class:"admin__system-card__item"},We=(0,a._)("span",null,"Пропуск",-1),Ge={class:"switch"},Je=(0,a._)("div",{class:"slider round"},null,-1),es={class:"admin__system-card__item"},ss=(0,a._)("span",null,"включен",-1),ts={class:"switch"},as=(0,a._)("div",{class:"slider round"},null,-1),ns={class:"admin__system-card__item"},is=(0,a._)("span",null,"Быстрый режим",-1),ls={class:"switch"},os=(0,a._)("div",{class:"slider round"},null,-1),_s={class:"admin__system-card__item"},ds=(0,a._)("span",null,"Количественный замер при положительном тесте",-1),cs={class:"switch"},rs=(0,a._)("div",{class:"slider round"},null,-1),ms={class:"admin__system-card animate__animated animate__fadeInUp d-3"},us=(0,a._)("div",{class:"admin__system-card__title"}," Тонометр ",-1),ps={class:"admin__system-card__item"},ys=(0,a._)("span",null,"Пропуск",-1),vs={class:"switch"},hs=(0,a._)("div",{class:"slider round"},null,-1),ws={class:"admin__system-card__item"},bs=(0,a._)("span",null,"включен",-1),gs={class:"switch"},fs=(0,a._)("div",{class:"slider round"},null,-1),ks={class:"admin__system-card animate__animated animate__fadeInUp d-4"},Us=(0,a._)("div",{class:"admin__system-card__title"}," Камера ",-1),Is={class:"admin__system-card__item"},Cs=(0,a._)("span",null,"Качество",-1),xs=["value"],Ds={class:"admin__system-card__item"},Vs=(0,a._)("span",null,"Видео",-1),qs={class:"switch"},zs=(0,a._)("div",{class:"slider round"},null,-1),$s={class:"admin__system-card__item"},Zs=(0,a._)("span",null,"Фото",-1),Ps={class:"switch"},Ls=(0,a._)("div",{class:"slider round"},null,-1),Ss={class:"admin__system-card__item"},Ts=(0,a._)("span",null,"Фото водителя",-1),js={class:"switch"},Fs=(0,a._)("div",{class:"slider round"},null,-1),Xs={class:"admin__system-card animate__animated animate__fadeInUp d-5"},As=(0,a._)("div",{class:"admin__system-card__title"}," Принтер ",-1),Hs={class:"admin__system-card__item"},Rs=(0,a._)("span",null,"Печать",-1),Ys={class:"switch"},Bs=(0,a._)("div",{class:"slider round"},null,-1),Es={class:"admin__system-card__item"},Ks=(0,a._)("span",null,"Количество",-1),Ns={class:"admin__system-card animate__animated animate__fadeInUp d-6"},Qs=(0,a._)("div",{class:"admin__system-card__title"}," Термометр ",-1),Ms={class:"admin__system-card__item"},Os=(0,a._)("span",null,"Пропуск",-1),Ws={class:"switch"},Gs=(0,a._)("div",{class:"slider round"},null,-1),Js={class:"admin__system-card__item"},et=(0,a._)("span",null,"включен",-1),st={class:"switch"},tt=(0,a._)("div",{class:"slider round"},null,-1),at={class:"admin__system-card block animate__animated animate__fadeInUp d-6"},nt={class:"admin__system-card__title"},it={class:"admin__system-card__item"},lt=(0,a._)("span",null,"Адрес",-1),ot={class:"admin__system-card__item"},_t=(0,a._)("span",null,"Токен",-1),dt={class:"admin__system-footer animate__animated animate__fadeInUp d-7"};function ct(e,s,t,i,l,o){return(0,a.wg)(),(0,a.iD)("div",be,[(0,a._)("div",ge,[fe,(0,a._)("div",ke,[Ue,(0,a._)("label",Ie,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[0]||(s[0]=e=>o.system.driver_info=e)},null,512),[[ie.e8,o.system.driver_info]]),Ce])]),(0,a._)("div",xe,[De,(0,a._)("label",Ve,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[1]||(s[1]=e=>o.system.type_ride=e)},null,512),[[ie.e8,o.system.type_ride]]),qe])]),(0,a._)("div",ze,[$e,(0,a._)("label",Ze,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[2]||(s[2]=e=>o.system.question_sleep=e)},null,512),[[ie.e8,o.system.question_sleep]]),Pe])]),(0,a._)("div",Le,[Se,(0,a._)("label",Te,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[3]||(s[3]=e=>o.system.question_helth=e)},null,512),[[ie.e8,o.system.question_helth]]),je])]),(0,a._)("div",Fe,[Xe,(0,a._)("label",Ae,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[4]||(s[4]=e=>o.system.manual_mode=e)},null,512),[[ie.e8,o.system.manual_mode]]),He])]),(0,a._)("div",Re,[Ye,(0,a._)("label",Be,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[5]||(s[5]=e=>o.system.auto_start=e)},null,512),[[ie.e8,o.system.auto_start]]),Ee])]),(0,a._)("div",Ke,[Ne,(0,a.wy)((0,a._)("input",{min:"1000",class:"medium",type:"number","onUpdate:modelValue":s[6]||(s[6]=e=>o.system.delay_before_retry_inspection=e)},null,512),[[ie.nr,o.system.delay_before_retry_inspection]])])]),(0,a._)("div",Qe,[Me,(0,a._)("div",Oe,[We,(0,a._)("label",Ge,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[7]||(s[7]=e=>o.system.alcometer_skip=e)},null,512),[[ie.e8,o.system.alcometer_skip]]),Je])]),(0,a._)("div",es,[ss,(0,a._)("label",ts,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[8]||(s[8]=e=>o.system.alcometer_visible=e)},null,512),[[ie.e8,o.system.alcometer_visible]]),as])]),(0,a._)("div",ns,[is,(0,a._)("label",ls,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[9]||(s[9]=e=>o.system.alcometer_fast=e)},null,512),[[ie.e8,o.system.alcometer_fast]]),os])]),(0,a._)("div",_s,[ds,(0,a._)("label",cs,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[10]||(s[10]=e=>o.system.alcometer_retry=e)},null,512),[[ie.e8,o.system.alcometer_retry]]),rs])])]),(0,a._)("div",ms,[us,(0,a._)("div",ps,[ys,(0,a._)("label",vs,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[11]||(s[11]=e=>o.system.tonometer_skip=e)},null,512),[[ie.e8,o.system.tonometer_skip]]),hs])]),(0,a._)("div",ws,[bs,(0,a._)("label",gs,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[12]||(s[12]=e=>o.system.tonometer_visible=e)},null,512),[[ie.e8,o.system.tonometer_visible]]),fs])])]),(0,a._)("div",ks,[Us,(0,a._)("div",Is,[Cs,(0,a.wy)((0,a._)("select",{"onUpdate:modelValue":s[13]||(s[13]=e=>o.system.camera_dimension=e)},[((0,a.wg)(!0),(0,a.iD)(a.HY,null,(0,a.Ko)(l.videoSizes,(e=>((0,a.wg)(),(0,a.iD)("option",{key:e,value:e},(0,n.zw)(e),9,xs)))),128))],512),[[ie.bM,o.system.camera_dimension]])]),(0,a._)("div",Ds,[Vs,(0,a._)("label",qs,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[14]||(s[14]=e=>o.system.camera_video=e)},null,512),[[ie.e8,o.system.camera_video]]),zs])]),(0,a._)("div",$s,[Zs,(0,a._)("label",Ps,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[15]||(s[15]=e=>o.system.camera_photo=e)},null,512),[[ie.e8,o.system.camera_photo]]),Ls])]),(0,a._)("div",Ss,[Ts,(0,a._)("label",js,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[16]||(s[16]=e=>o.system.driver_photo=e)},null,512),[[ie.e8,o.system.driver_photo]]),Fs])])]),(0,a._)("div",Xs,[As,(0,a._)("div",Hs,[Rs,(0,a._)("label",Ys,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[17]||(s[17]=e=>o.system.printer_write=e)},null,512),[[ie.e8,o.system.printer_write]]),Bs])]),(0,a._)("div",Es,[Ks,(0,a.wy)((0,a._)("input",{min:"1",class:"small",type:"number","onUpdate:modelValue":s[18]||(s[18]=e=>o.system.print_count=e)},null,512),[[ie.nr,o.system.print_count]])])]),(0,a._)("div",Ns,[Qs,(0,a._)("div",Ms,[Os,(0,a._)("label",Ws,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[19]||(s[19]=e=>o.system.thermometer_skip=e)},null,512),[[ie.e8,o.system.thermometer_skip]]),Gs])]),(0,a._)("div",Js,[et,(0,a._)("label",st,[(0,a.wy)((0,a._)("input",{type:"checkbox","onUpdate:modelValue":s[20]||(s[20]=e=>o.system.thermometer_visible=e)},null,512),[[ie.e8,o.system.thermometer_visible]]),tt])])]),(0,a._)("div",at,[(0,a._)("div",nt,[(0,a.Uk)(" Настройки API "),(0,a._)("div",{class:(0,n.C_)(["connection-status",{active:o.connection}])},[(0,a._)("span",null,(0,n.zw)(o.connection),1)],2)]),(0,a._)("div",it,[lt,(0,a.wy)((0,a._)("input",{"onUpdate:modelValue":s[21]||(s[21]=e=>o.config.url=e)},null,512),[[ie.nr,o.config.url]])]),(0,a._)("div",ot,[_t,(0,a.wy)((0,a._)("input",{"onUpdate:modelValue":s[22]||(s[22]=e=>o.config.token=e)},null,512),[[ie.nr,o.config.token]])])]),(0,a._)("div",dt,[(0,a._)("button",{onClick:s[23]||(s[23]=(...e)=>o.save&&o.save(...e)),class:"btn blue"},"Сохранить")])])}var rt=t(6509),mt={data(){return{toast:(0,se.pm)(),videoSizes:[]}},async mounted(){this.videoSizes=await(0,P.TQ)()},methods:{async save(){await(0,ee.hR)(this.system),await(0,ee.XP)(this.config.url,this.config.token),this.$store.state.point=await(0,rt.F8)(),this.toast.success("Настройки сохранены")}},computed:{system(){return this.$store.state.config?.system||{}},config(){return this.$store.state.config?.main||{}},connection(){return this.$store.state.connection||!1}}};const ut=(0,R.Z)(mt,[["render",ct]]);var pt=ut;const yt={class:"admin__logo"},vt={class:"input-file"},ht=(0,a._)("span",null,"Выберите файл",-1),wt={key:0,class:"preview-image"},bt=["src"];function gt(e,s,t,n,i,l){return(0,a.wg)(),(0,a.iD)("div",yt,[(0,a._)("label",vt,[(0,a._)("input",{type:"file",onChange:s[0]||(s[0]=(...e)=>l.onFileChange&&l.onFileChange(...e))},null,32),ht]),l.logo?((0,a.wg)(),(0,a.iD)("div",wt,[(0,a._)("img",{src:l.logo},null,8,bt),(0,a._)("button",{onClick:s[1]||(s[1]=(...e)=>l.deleteLogo&&l.deleteLogo(...e)),class:"btn small"},"Удалить")])):(0,a.kq)("",!0)])}var ft={data(){return{toast:(0,se.pm)()}},methods:{async onFileChange(e){const s=e.target.files[0],t=await this.getBase64(s);this.$store.state.config.main.logo=t,await(0,rt.Sx)(t),this.toast.success("Логотип успешно установлен")},getBase64(e){return new Promise(((s,t)=>{const a=new FileReader;a.readAsDataURL(e),a.onload=()=>{s(a.result)},a.onerror=e=>{console.log("Error: ",e)}}))},async deleteLogo(){this.$store.state.config.main.logo="",await(0,rt.Sx)(""),this.toast.success("Логотип успешно удален")}},computed:{logo(){return this.$store.state.config?.main?.logo}}};const kt=(0,R.Z)(ft,[["render",gt]]);var Ut=kt,It={components:{Testing:B,Tonometer:ne,Password:we,System:pt,Logo:Ut},data(){return{selected:"system"}}};const Ct=(0,R.Z)(It,[["render",o]]);var xt=Ct}}]);
-//# sourceMappingURL=975.faf22561.js.map
