+++ src/main/resources/static/js/app.3e91c0e7.js
@@ -1,0 +1,2 @@
+(function(){"use strict";var e={6700:function(e,t,n){n.d(t,{_L:function(){return s},bP:function(){return a},pe:function(){return o},qL:function(){return c}});var i=n(1857);async function o(){return await axios.post("device/alcometer").then((({data:e})=>e)).catch(i.i)}async function s(){return await axios.post("device/alcometer/close").then((({data:e})=>e)).catch(i.i)}async function r(){return u("fast")}async function a(){return u("slow")}async function c(e){return e?r():e?void 0:a()}async function u(e){return await axios.post("device/alcometer/change-mode",{mode:e}).then((({data:e})=>e)).catch(i.i)}},6509:function(e,t,n){n.d(t,{Ch:function(){return y},F8:function(){return h},Jw:function(){return p},L$:function(){return c},Ni:function(){return f},S1:function(){return u},Sx:function(){return _},_u:function(){return m},oP:function(){return g},oX:function(){return d},rd:function(){return l},sl:function(){return b},vs:function(){return w},xH:function(){return k},xv:function(){return v}});var i=n(7625),o=n(4161),s=n(1857),r=n(4677);const a=(0,r.pm)();async function c(e){return await o.Z.post(`inspection/${e}`).then((({data:e})=>e))}async function u(e,t){return await o.Z.post("api/photo",{driver_id:e,photo:t}).then((({data:e})=>e))}async function l(){return await o.Z.post("api/photo/stop").then((({data:e})=>e))}async function m(e=i.Z.state.inspection){return i.Z.state.config?.main?.selected_medic?.id&&(e.user_id=i.Z.state.config.main.selected_medic.id),await o.Z.post("inspection/save",e).then((({data:e})=>(console.log(e),e))).catch(s.i)}async function p(){return await o.Z.post("inspection/print").then((({data:e})=>e)).catch(s.i)}async function d(e){return await o.Z.post("api/check",{address:e}).then((({data:e})=>e)).catch((e=>{e.response||(i.Z.state.loseConnect=!0)}))}async function h(){return await o.Z.get("api/pv").then((({data:e})=>e)).catch((e=>{console.log(e)}))}async function f(){let e=await o.Z.get("api/verification").then((({data:e})=>e)).catch((e=>{console.log(e)})),t=e.date_check??"",n=e.serial_number??"";return{dateInspection:t,serialNumberTerminal:n}}async function g(){return await o.Z.get("api/medics").then((({data:e})=>e)).catch((e=>{console.log(e)}))}async function v(){o.Z.post("exit").then((({data:e})=>{})).catch((e=>{console.log(e)})),window.location.reload()}async function _(e){return await o.Z.post("api/logo",{logo:e}).then((({data:e})=>e)).catch((e=>{console.log(e)}))}async function b(e){return await o.Z.post("api/medic",e).then((({data:e})=>e)).catch((e=>{console.log(e)}))}async function y(e,t){return await o.Z.post("api/driver/phone/save",{phone_number:e,driver_id:t}).then((({data:e})=>(a.success("Личный номер телефона сохранен"),e))).catch((e=>{console.log(e),a.error("Ошибка сохранения номера телефона. Повторите попытку или обратитесь к администратору")}))}async function w(e){return await o.Z.post(`api/${e}/inspections/`).then((({data:e})=>e)).catch((e=>{console.log(e)}))}async function k(e){return await o.Z.post("/device/printer/inspection",e).then((({data:e})=>e)).catch((e=>e))}},1857:function(e,t,n){n.d(t,{i:function(){return r}});var i=n(7625),o=n(4677);const s=(0,o.pm)();function r(e){const t=e.response?.data;if(t?.message)s.error(t.message);else if(e.response)switch(e?.response?.status){case 400:s.error("Ошибка авторизации запроса");break;case 500:s.error("Ошибка сервера");break;default:s.error("Неизвестная ошибка запроса")}else i.Z.state.loseConnect=!0;console.log(e)}},4457:function(e,t,n){n.d(t,{Tc:function(){return l},XP:function(){return u},hR:function(){return c},ri:function(){return r},vW:function(){return a}});var i=n(7625),o=n(4161),s=n(6509);async function r(){await o.Z.post("/setting/load").then((({data:e})=>{i.Z.state.config=e,e?.system?.cursor&&document.documentElement.classList.remove("disable-mouse")})).catch((e=>{console.log(e)})),i.Z.state.point=await(0,s.F8)(),i.Z.state.verification=await(0,s.Ni)()}async function a(e){await o.Z.post("/setting/password",{password:e}).catch((e=>{console.log(e)}))}async function c(e){await o.Z.post("/setting/system",e).catch((e=>{console.log(e)}))}async function u(e,t){await o.Z.post("/setting/api",{address:e,token:t}).catch((e=>{console.log(e)}))}async function l(e){await o.Z.post("/setting/tonometer",{address:e}).catch((e=>{console.log(e)}))}},6026:function(e,t,n){n(7658);var i=n(9242),o=n(3396),s=n(7139);const r={class:"footer"},a={class:"footer__point-version"},c=(0,o._)("span",null,"Версия: 2.5.11",-1),u={class:"footer__serial-number_date-notification"},l={class:"date-notification"},m={class:"serial-number"};function p(e,t,n,p,d,h){const f=(0,o.up)("navigation"),g=(0,o.up)("error-modal"),v=(0,o.up)("router-view");return(0,o.wg)(),(0,o.iD)(o.HY,null,[(0,o.Wm)(f),(0,o.Wm)(g),(0,o.Wm)(i.uT,{name:"fade"},{default:(0,o.w5)((()=>[(0,o.Wm)(v)])),_:1}),(0,o._)("div",r,[(0,o._)("div",a,[(0,o.Uk)((0,s.zw)(h.point||"Незвестный пункт выпуска")+" ",1),c])]),(0,o._)("div",u,[(0,o._)("div",null,[(0,o._)("span",l,(0,s.zw)(h.dateVerification||""),1)]),(0,o._)("span",m,(0,s.zw)(h.serialNumber?"S/N "+h.serialNumber:""),1)])],64)}var d=n.p+"img/logo.957d00d6.png";const h={class:"nav"},f={class:"nav__logo animate__animated animate__fadeInDown"},g=["src"],v={key:1,src:d,alt:""},_={key:0,class:"nav__info animate__animated animate__fadeInDown"},b={key:1,class:"nav__buttons"},y={key:2,class:"nav__buttons"},w={key:3,class:"nav__buttons"},k={key:0,class:"step-progress animate__animated animate__fadeInDown"};function Z(e,t,n,i,r,a){const c=(0,o.up)("admin-navigation"),u=(0,o.up)("login-navigation"),l=(0,o.up)("InspectionProgressLine");return(0,o.wg)(),(0,o.iD)(o.HY,null,[(0,o._)("div",h,[(0,o._)("div",f,[a.logo?((0,o.wg)(),(0,o.iD)("img",{key:0,src:a.logo,alt:""},null,8,g)):((0,o.wg)(),(0,o.iD)("img",v))]),a.isInspectionPage?((0,o.wg)(),(0,o.iD)("div",_,[(0,o.Uk)((0,s.zw)(a.inspection.driver_fio)+" ",1),(0,o._)("span",null,(0,s.zw)(a.inspection.driver_id),1)])):(0,o.kq)("",!0),a.needButtonRedirectOnHomePage?((0,o.wg)(),(0,o.iD)("div",b,[(0,o._)("button",{onClick:t[0]||(t[0]=t=>e.$router.push("/")),class:"btn opacity blue animate__animated animate__fadeInDown d-1"},"В начало ")])):(0,o.kq)("",!0),a.isAdminPage?((0,o.wg)(),(0,o.iD)("div",y,[(0,o.Wm)(c,{onChangeVisibleMouseCursor:t[1]||(t[1]=e=>a.changeVisibleMouseCursor()),onChangeModeFullScreen:t[2]||(t[2]=e=>a.changeModeFullScreen()),onLogout:t[3]||(t[3]=e=>a.logout())})])):(0,o.kq)("",!0),a.isHomePage?((0,o.wg)(),(0,o.iD)("div",w,[(0,o.Wm)(u,{onSelectMedic:t[4]||(t[4]=t=>e.$store.state.selectingMedic=!0),onGetHelp:t[5]||(t[5]=t=>e.$router.push("/help")),onGetLastInspection:t[6]||(t[6]=t=>e.$router.push("/print/index")),onSettings:t[7]||(t[7]=t=>e.$router.push("/login"))})])):(0,o.kq)("",!0)]),a.isInspectionPage?((0,o.wg)(),(0,o.iD)("div",k,[(0,o.Wm)(l)])):(0,o.kq)("",!0)],64)}var P=n(6509);const C=(0,o._)("i",{class:"ri-cursor-fill"},null,-1),N=[C],x=(0,o._)("i",{class:"ri-fullscreen-line"},null,-1),S=[x];function D(e,t,n,i,s,r){return(0,o.wg)(),(0,o.iD)(o.HY,null,[(0,o._)("button",{onClick:t[0]||(t[0]=e=>r.changeVisibleMouseCursor()),class:"btn icon animate__animated animate__fadeInDown d-1",style:{padding:"5px"}},N),(0,o._)("button",{onClick:t[1]||(t[1]=e=>r.changeModeFullScreen()),class:"btn icon animate__animated animate__fadeInDown d-1",style:{padding:"5px"}},S),(0,o._)("button",{onClick:t[2]||(t[2]=e=>r.logout()),class:"btn opacity blue animate__animated animate__fadeInDown d-2"},"Выйти")],64)}var F={methods:{changeVisibleMouseCursor(){this.$emit("changeVisibleMouseCursor")},changeModeFullScreen(){this.$emit("changeModeFullScreen")},logout(){this.$emit("logout")}}},q=n(89);const I=(0,q.Z)(F,[["render",D]]);var O=I;const $=(0,o._)("i",{class:"ri-tools-fill"},null,-1),E=[$];function L(e,t,n,i,s,r){return(0,o.wg)(),(0,o.iD)(o.HY,null,[(0,o._)("button",{onClick:t[0]||(t[0]=e=>r.selectMedic()),class:"btn opacity animate__animated animate__fadeInDown"},"Мед работник "),(0,o._)("button",{onClick:t[1]||(t[1]=e=>r.getHelp()),class:"btn opacity animate__animated animate__fadeInDown d-2"},"Помощь "),(0,o._)("button",{onClick:t[2]||(t[2]=e=>r.getLastInspection()),class:"btn opacity animate__animated animate__fadeInDown d-2"},"Повторная печать "),(0,o._)("button",{onClick:t[3]||(t[3]=e=>r.settings()),class:"btn icon animate__animated animate__fadeInDown d-3"},E)],64)}var M={methods:{selectMedic(){this.$emit("selectMedic")},getHelp(){this.$emit("getHelp")},settings(){this.$emit("settings")},getLastInspection(){this.$emit("getLastInspection")}}};const R=(0,q.Z)(M,[["render",L]]);var H=R;const T=(0,o._)("i",{class:"ri-check-fill"},null,-1),j=[T],z=(0,o._)("i",{class:"ri-check-fill"},null,-1),A=[z],J=(0,o._)("i",{class:"ri-check-fill"},null,-1),V=[J],W=(0,o._)("i",{class:"ri-check-fill"},null,-1),U=[W],Y=(0,o._)("i",{class:"ri-check-fill"},null,-1),B=[Y],X=(0,o._)("i",{class:"ri-check-fill"},null,-1),G=[X],K=(0,o._)("i",{class:"ri-check-fill"},null,-1),Q=[K],ee=(0,o._)("i",{class:"ri-check-fill"},null,-1),te=[ee],ne=(0,o._)("i",{class:"ri-check-fill"},null,-1),ie=[ne];function oe(e,t,n,i,r,a){return(0,o.wg)(),(0,o.iD)(o.HY,null,[JSON.parse(a.system.driver_info)?((0,o.wg)(),(0,o.iD)("div",{key:0,class:(0,s.C_)(["step-progress__item",{active:a.pageNumber>=1}])},j,2)):(0,o.kq)("",!0),JSON.parse(a.system.type_ride)?((0,o.wg)(),(0,o.iD)("div",{key:1,class:(0,s.C_)(["step-progress__item",{active:a.pageNumber>=2}])},A,2)):(0,o.kq)("",!0),JSON.parse(a.system.type_ride)?((0,o.wg)(),(0,o.iD)("div",{key:2,class:(0,s.C_)(["step-progress__item",{active:a.pageNumber>=3}])},V,2)):(0,o.kq)("",!0),JSON.parse(a.system.tonometer_visible)?((0,o.wg)(),(0,o.iD)("div",{key:3,class:(0,s.C_)(["step-progress__item",{active:a.pageNumber>=4}])},U,2)):(0,o.kq)("",!0),JSON.parse(a.system.thermometer_visible)?((0,o.wg)(),(0,o.iD)("div",{key:4,class:(0,s.C_)(["step-progress__item",{active:a.pageNumber>=5}])},B,2)):(0,o.kq)("",!0),JSON.parse(a.system.alcometer_visible)?((0,o.wg)(),(0,o.iD)("div",{key:5,class:(0,s.C_)(["step-progress__item",{active:a.pageNumber>=6}])},G,2)):(0,o.kq)("",!0),JSON.parse(a.system.question_sleep)?((0,o.wg)(),(0,o.iD)("div",{key:6,class:(0,s.C_)(["step-progress__item",{active:a.pageNumber>=7}])},Q,2)):(0,o.kq)("",!0),JSON.parse(a.system.question_helth)?((0,o.wg)(),(0,o.iD)("div",{key:7,class:(0,s.C_)(["step-progress__item",{active:a.pageNumber>=8}])},te,2)):(0,o.kq)("",!0),(0,o._)("div",{class:(0,s.C_)(["step-progress__item",{active:a.pageNumber>=9}])},ie,2)],64)}var se={computed:{system(){return this.$store.state.config?.system||{}},pageNumber(){return this.$route.meta?.number||0}}};const re=(0,q.Z)(se,[["render",oe]]);var ae=re,ce={components:{InspectionProgressLine:ae,LoginNavigation:H,AdminNavigation:O},data(){return{}},methods:{close:P.xv,changeVisibleMouseCursor(){document.documentElement.classList.contains("disable-mouse")?(document.documentElement.classList.remove("disable-mouse"),this.system.cursor=!0):(document.documentElement.classList.add("disable-mouse"),this.system.cursor=!1)},logout(){this.$store.state.admin=!1,this.$router.push("/")},changeModeFullScreen(){const e=document.fullscreenElement||document.webkitFullscreenElement||document.mozFullScreenElement||document.msFullscreenElement,t=document.documentElement;e?document.exitFullscreen?document.exitFullscreen():document.webkitExitFullscreen?document.webkitExitFullscreen():document.mozCancelFullScreen?document.mozCancelFullScreen():document.msExitFullscreen&&document.msExitFullscreen():t.requestFullscreen?t.requestFullscreen():t.mozRequestFullScreen?t.mozRequestFullScreen():t.webkitRequestFullScreen?t.webkitRequestFullScreen():t.msRequestFullscreen&&t.msRequestFullscreen()}},computed:{needButtonRedirectOnHomePage(){return this.isHelpPage||this.isInspectionPage||this.isLoginPage||this.isPrintPage||this.isSetNumberPhonePage},isHomePage(){return!(this.isHelpPage||this.isInspectionPage||this.isLoginPage||this.isAdminPage||this.isPrintPage||this.isSetNumberPhonePage)},isHelpPage(){return this.currentRouter.includes("/help")},isLoginPage(){return this.currentRouter.includes("/login")},isAdminPage(){return this.currentRouter.includes("/admin")},isInspectionPage(){return this.currentRouter.includes("/step")},isPrintPage(){return this.currentRouter.includes("/print")},isSetNumberPhonePage(){return this.currentRouter.includes("/number-phone/add/")},currentRouter(){return this.$route.path},inspection(){return this.$store.state.inspection},logo(){return this.$store.state.config?.main?.logo}}};const ue=(0,q.Z)(ce,[["render",Z]]);var le=ue;const me={key:0,class:"alert-container"},pe=(0,o._)("div",{class:"alert"},[(0,o._)("h2",null,"Потеряна связь"),(0,o.Uk)(" Пожалуйста, перезапустите СДПО ")],-1),de=[pe];function he(e,t,n,i,s,r){return r.loseConnect?((0,o.wg)(),(0,o.iD)("div",me,de)):(0,o.kq)("",!0)}var fe={data(){},computed:{loseConnect(){return this.$store.state.loseConnect}}};const ge=(0,q.Z)(fe,[["render",he]]);var ve=ge,_e={name:"App",components:{Navigation:le,ErrorModal:ve},computed:{point(){return this.$store.state.point},dateVerification(){return this.$store.state.verification.dateInspection},serialNumber(){return this.$store.state.verification.serialNumberTerminal}}};const be=(0,q.Z)(_e,[["render",p]]);var ye=be,we=n(7625),ke=n(4161),Ze=n(2483);const Pe=[{path:"",name:"home",component:()=>n.e(330).then(n.bind(n,5330))},{path:"/help",component:()=>n.e(988).then(n.bind(n,6988))},{path:"/login",component:()=>n.e(663).then(n.bind(n,9663))},{path:"/print/index",component:()=>n.e(733).then(n.bind(n,733))},{path:"/admin",component:()=>n.e(975).then(n.bind(n,1975))},{path:"/number-phone/add",component:()=>n.e(702).then(n.bind(n,3702))},{path:"/step/driver",component:()=>n.e(625).then(n.bind(n,5625)),name:"step-driver",meta:{visible:"driver_info",next:"step-photo",prev:"home",number:1}},{path:"/step/photo",component:()=>n.e(57).then(n.bind(n,8057)),name:"step-photo",meta:{visible:"driver_photo",next:"step-ride",prev:"step-driver",number:2}},{path:"/step/ride",component:()=>n.e(267).then(n.bind(n,267)),name:"step-ride",meta:{visible:"type_ride",next:"step-tonometer",prev:"step-driver",number:3}},{path:"/step/tonometer",component:()=>n.e(286).then(n.bind(n,3286)),name:"step-tonometer",meta:{visible:"tonometer_visible",next:"step-thermometer",prev:"step-ride",number:4}},{path:"/step/thermometer",component:()=>n.e(381).then(n.bind(n,3524)),name:"step-thermometer",meta:{visible:"thermometer_visible",next:"step-alcometer",prev:"step-thermometer",number:5}},{path:"/step/alcometer",component:()=>n.e(847).then(n.bind(n,5229)),name:"step-alcometer",meta:{visible:"alcometer_visible",next:"step-sleep",prev:"step-alcometer",number:6}},{path:"/step/sleep",component:()=>n.e(449).then(n.bind(n,9449)),name:"step-sleep",meta:{visible:"question_sleep",next:"step-helth",prev:"step-alcometer",number:7}},{path:"/step/helth",component:()=>n.e(787).then(n.bind(n,9787)),name:"step-helth",meta:{visible:"question_helth",next:"step-result",prev:"step-sleep",number:8}},{path:"/step/result",component:()=>n.e(885).then(n.bind(n,5606)),name:"step-result",meta:{next:"home",prev:"step-result",number:9}},{path:"/step/retry",component:()=>n.e(885).then(n.bind(n,5606)),name:"step-retry",meta:{next:"step-driver",prev:"step-result",number:0}}];var Ce=n(4677),Ne=(n(343),n(4457)),xe=n(6700),Se=n(7190);ke.Z.defaults.baseURL="http://localhost:8080/",window.axios=ke.Z,(0,Ne.ri)();const De=(0,Ze.p7)({history:(0,Ze.PO)(),routes:Pe,mode:"history",scrollBehavior(e,t,n){return n||{top:0,left:0}}});De.beforeEach(((e,t,n)=>{if("step-alcometer"===t.name&&(0,xe._L)(),"step-retry"===e.name)return t.meta.number=0,De.push({name:e.meta.next});if(e.path.includes("/step/")){if(!we.Z.state.inspection?.driver_id)return De.push({name:"home"});if(e.meta?.visible&&we.Z.state.config?.system&&!JSON.parse(we.Z.state.config.system[e.meta.visible]))return t.meta?.number>e.meta.number?De.push({name:e.meta.prev}):De.push({name:e.meta.next})}return e.path.includes("/admin/")&&!we.Z.state.admin?De.push({name:"home"}):("home"===e.name&&(we.Z.state.inspection={}),n())})),setInterval((async()=>{we.Z.state.config?.main?.url&&(we.Z.state.connection=await(0,P.oX)(we.Z.state.config.main.url))}),2e3),(0,i.ri)(ye).use(we.Z).use(Ce.ZP,{position:"top-center",timeout:4500}).use(De).component("VueDatePicker",Se.Z).mount("#app")},7625:function(e,t,n){n(1703);var i=n(65),o=n(4677);const s=(0,o.pm)(),r=(0,i.MT)({state(){return{inspection:{},driver:{},config:{},admin:!1,connection:!1,point:"",verification:{serialNumberTerminal:"",dateInspection:""},selectingMedic:!1,loseConnect:!1}},getters:{url:e=>{if(e.config?.main?.url)return e.config.main.url;throw s.error("Ошибка конфигурации. URL адрес не указан."),new Error("Error configuration API url")}}});t["Z"]=r}},t={};function n(i){var o=t[i];if(void 0!==o)return o.exports;var s=t[i]={exports:{}};return e[i](s,s.exports,n),s.exports}n.m=e,function(){var e=[];n.O=function(t,i,o,s){if(!i){var r=1/0;for(l=0;l<e.length;l++){i=e[l][0],o=e[l][1],s=e[l][2];for(var a=!0,c=0;c<i.length;c++)(!1&s||r>=s)&&Object.keys(n.O).every((function(e){return n.O[e](i[c])}))?i.splice(c--,1):(a=!1,s<r&&(r=s));if(a){e.splice(l--,1);var u=o();void 0!==u&&(t=u)}}return t}s=s||0;for(var l=e.length;l>0&&e[l-1][2]>s;l--)e[l]=e[l-1];e[l]=[i,o,s]}}(),function(){n.n=function(e){var t=e&&e.__esModule?function(){return e["default"]}:function(){return e};return n.d(t,{a:t}),t}}(),function(){n.d=function(e,t){for(var i in t)n.o(t,i)&&!n.o(e,i)&&Object.defineProperty(e,i,{enumerable:!0,get:t[i]})}}(),function(){n.f={},n.e=function(e){return Promise.all(Object.keys(n.f).reduce((function(t,i){return n.f[i](e,t),t}),[]))}}(),function(){n.u=function(e){return"js/"+e+"."+{57:"d3ec11d8",267:"a461acce",286:"a327899b",330:"cc3e1306",381:"acac5138",449:"05dbda15",625:"9aee366c",663:"ea8d0713",702:"5c893480",733:"fcc3d413",787:"9c8989ae",847:"ac14933c",885:"ae66229d",975:"14e4eb48",988:"4313f46b"}[e]+".js"}}(),function(){n.miniCssF=function(e){}}(),function(){n.g=function(){if("object"===typeof globalThis)return globalThis;try{return this||new Function("return this")()}catch(e){if("object"===typeof window)return window}}()}(),function(){n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)}}(),function(){var e={},t="sdpo-frontend:";n.l=function(i,o,s,r){if(e[i])e[i].push(o);else{var a,c;if(void 0!==s)for(var u=document.getElementsByTagName("script"),l=0;l<u.length;l++){var m=u[l];if(m.getAttribute("src")==i||m.getAttribute("data-webpack")==t+s){a=m;break}}a||(c=!0,a=document.createElement("script"),a.charset="utf-8",a.timeout=120,n.nc&&a.setAttribute("nonce",n.nc),a.setAttribute("data-webpack",t+s),a.src=i),e[i]=[o];var p=function(t,n){a.onerror=a.onload=null,clearTimeout(d);var o=e[i];if(delete e[i],a.parentNode&&a.parentNode.removeChild(a),o&&o.forEach((function(e){return e(n)})),t)return t(n)},d=setTimeout(p.bind(null,void 0,{type:"timeout",target:a}),12e4);a.onerror=p.bind(null,a.onerror),a.onload=p.bind(null,a.onload),c&&document.head.appendChild(a)}}}(),function(){n.r=function(e){"undefined"!==typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})}}(),function(){n.p="/"}(),function(){var e={143:0};n.f.j=function(t,i){var o=n.o(e,t)?e[t]:void 0;if(0!==o)if(o)i.push(o[2]);else{var s=new Promise((function(n,i){o=e[t]=[n,i]}));i.push(o[2]=s);var r=n.p+n.u(t),a=new Error,c=function(i){if(n.o(e,t)&&(o=e[t],0!==o&&(e[t]=void 0),o)){var s=i&&("load"===i.type?"missing":i.type),r=i&&i.target&&i.target.src;a.message="Loading chunk "+t+" failed.\n("+s+": "+r+")",a.name="ChunkLoadError",a.type=s,a.request=r,o[1](a)}};n.l(r,c,"chunk-"+t,t)}},n.O.j=function(t){return 0===e[t]};var t=function(t,i){var o,s,r=i[0],a=i[1],c=i[2],u=0;if(r.some((function(t){return 0!==e[t]}))){for(o in a)n.o(a,o)&&(n.m[o]=a[o]);if(c)var l=c(n)}for(t&&t(i);u<r.length;u++)s=r[u],n.o(e,s)&&e[s]&&e[s][0](),e[s]=0;return n.O(l)},i=self["webpackChunksdpo_frontend"]=self["webpackChunksdpo_frontend"]||[];i.forEach(t.bind(null,0)),i.push=t.bind(null,i.push.bind(i))}();var i=n.O(void 0,[998],(function(){return n(6026)}));i=n.O(i)})();
+//# sourceMappingURL=app.3e91c0e7.js.map
